code rework
